requirements.txt
streamlit
openai
pypdf
faiss-cpu
sentence-transformers
numpy
soundfile
whisper
torch
gtts
